+++ bbsApp/controllers/userController.js
@@ -1,12 +1,6 @@
 const userModel = require('../models/usersModel');
 
 module.exports = {
-  goAllPosts(req, res) {
-    res.render('users/allPosts', { title: `Everyone's Posts`, data: res.user, posts: res.posts} );
-  },
-  goMyPost(req, res) {
-    res.render('users/myPost', { title: 'Post new post', data: res.user} );
-  },
   async getUserName(req, res, next) {
     await userModel.compareEmail(req, res);
     next();
@@ -15,8 +9,5 @@ module.exports = {
     await userModel.insertUserData(req, res);
     next();
   },
-  goEditPost(req, res) {
-    res.render('users/editPost', { title: 'Edit Post', data: res.user, post: res.post} );
-  }
 
 }
