+++ bbsApp/modules/authenticator.js
@@ -5,7 +5,7 @@ const jwtConfig = {
   secret: 'secret_key',
   options: {
     algorithm: 'HS256',
-    expiresIn: '10s'
+    expiresIn: '1m'
   }
 }
