+++ bbsApp/routes/index.js
@@ -2,8 +2,8 @@ const express = require('express');
 const authController = require('../controllers/authController');
 const router = express.Router();
 const authRoute = require('../routes/authRoute');
-const userRoute = require('../routes/userRoute');
+const postRoute = require('../routes/postRoute');
 
 router.get('/', authController.goLogin);
 
-module.exports = [router, authRoute, userRoute]
+module.exports = [router, authRoute, postRoute]
