+++ bbsApp/views/components/navbar.ejs
@@ -9,10 +9,10 @@
     <div class="collapse navbar-collapse" id="navbarSupportedContent">
       <ul class="navbar-nav ms-auto mb-2 mb-lg-0">
         <li class="nav-item">
-          <a class="nav-link" href="/users/allPosts">みんなの投稿</a>
+          <a class="nav-link" href="/posts/allPosts">みんなの投稿</a>
         </li>
         <li class="nav-item">
-          <a class="nav-link" href="/users/myPost">投稿を作成</a>
+          <a class="nav-link" href="/posts/makePost">投稿を作成</a>
         </li>
         <li class="nav-item">
           <a class="nav-link" href="/auth/login">Login</a>
