+++ bbsApp/views/posts/makePost.ejs
@@ -1,7 +1,7 @@
 <%- include('../components/header'); %>
 <%- include('../components/navbar_logged_in'); %>
 <div class="container mx-5">
-  <form action="../users/createPost" method="post" class="row needs-validation" novalidate>
+  <form action="../posts/createPost" method="POST" class="row needs-validation" novalidate>
     <legend class="mb-3">新規投稿</legend>
     <div class="mb-3">
       <label for="title" class="form-label">タイトル</label>
